# Extend Termination Checking with Dummy Arguments