--- a/test/Succeed/Issue7693a.agda
+++ b/test/Succeed/Issue7693a.agda
@@ -1,14 +1,18 @@
 open import Agda.Builtin.Nat using (zero; suc; _+_) renaming (Nat to ℕ)
 
 mutual
-  f : (n : ℕ) → ℕ
-  f (suc n) = g zero (f n)
-  f _ = 42
 
-  g : (m k : ℕ) → ℕ
-  g m (suc k) = g m k
-  g (suc zero) k = g zero (f 42)
-  g _ _ = 42
+  id : (n : ℕ) → ℕ
+  id (suc n) = wks zero (id n)
+  id _ = 42
+
+  wk : (q k : ℕ) → ℕ
+  wk (suc zero) k = wks zero (id 42)
+  wk _ _ = 42
+
+  wks : (q k : ℕ) → ℕ
+  wks q (suc k) = wk q k
+  wks _ _ = 42
 
 -- Used to be a termination error
 -- *Arguably* should succeed