--- a/test/Succeed/Issue7693b.agda
+++ b/test/Succeed/Issue7693b.agda
@@ -1,21 +1,50 @@
-open import Agda.Builtin.Nat using (zero; suc; _+_) renaming (Nat to ℕ)
+open import Agda.Builtin.Nat using () renaming (Nat to ℕ; zero to ze; suc to su)
+open import Agda.Builtin.Equality using (_≡_; refl)
+
+module _ where
+
+data Sort : Set where
+  V   : Sort
+  T>V : ∀ q → q ≡ V → Sort
+
+pattern T = T>V V refl
+
+variable
+  q r s : Sort
+
+_⊔_ : Sort → Sort → Sort
+V ⊔ r = r
+T ⊔ r = T
+
+variable
+  Γ Δ : ℕ
+
+data Tm[_] : Sort → ℕ → Set where
+  ƛ_  : Tm[ T ] (su Γ) → Tm[ T ] Γ
+
+data Tms[_] (q : Sort) (Δ : ℕ) : ℕ → Set where
+  _,_ : Tms[ q ] Δ Γ → Tm[ q ] Δ → Tms[ q ] Δ (su Γ)
+
+postulate
+  var0  : ∀ q → Tm[ q ] (su Γ)
+  ε     : Tms[ q ] Δ ze
+
+data Unit : Set where
+  unit : Unit
 
 mutual
-  id : (n : ℕ) → ℕ
-  id (suc n) = wks zero n n (id n)
-  id _ = 42
+  id : ∀ Γ → Tms[ V ] Γ Γ
+  id (su Γ) = wks V Γ Γ (id Γ) , var0 V
+  id ze     = ε
 
-  wk : (q n k : ℕ) → ℕ
-  wk (suc zero) n k = sub zero n (suc n) k (wks zero n n (id n))
-  wk _ _ _ = 42
+  wks : ∀ q Δ Γ → Tms[ q ] Δ Γ → Tms[ q ] (su Δ) Γ
+  wks q Δ (su Γ) (δ , x) = wks q Δ Γ δ , wk q Δ x
 
-  wks : (q m n h : ℕ) → ℕ
-  wks q m (suc n) (suc h) = wk q m h + wks q m n h
-  wks _ _ _ _ = 42
+  wk : ∀ q Γ → Tm[ q ] Γ → Tm[ q ] (su Γ)
+  wk T Γ t = sub T V Γ (su Γ) t (wks V Γ Γ (id Γ))
 
-  sub : (q n m k l : ℕ) → ℕ
-  sub q n m (suc k) l = sub q (suc n) (suc m) k (suc (wks q m n l))
-  sub _ _ _ _ _ = 42
+  sub : ∀ q r Γ Δ →  Tm[ q ] Γ → Tms[ r ] Δ Γ → Tm[ q ⊔ r ] Δ
+  sub q r Γ Δ (ƛ t) δ = ƛ (sub T r (su Γ) (su Δ) t (wks r Δ Γ δ , var0 r))
 
 -- Used to be a termination error
 -- *Arguably* should succeed