Implemented the termination checking extension suggested by @andreasabel in https://github.com/agda/agda/issues/7693#issuecomment-2621495501.

Really naive implementation that just adds a row and column of `=`s to every call matrix. This will obviously increase the size of all termination matrices, so probably has a performance impact. I imagine there are much smarter way to do this...