@NathanielB123 2025-01-29
Focussing on the standard library fail (`Tactic.RingSolver.Core.Polynomial.Base`), the matrices all look as I would expect (one extra row/column of `=`s between the guardedness stuff and the main call matrix) so I'm not sure what has gone wrong. Maybe there is some code somewhere that assumes indices of matrix entries correspond to arguments?

[Success.txt](https://github.com/user-attachments/files/18592618/Success.txt)
[Failure.txt](https://github.com/user-attachments/files/18592617/Failure.txt)