@NathanielB123 2025-01-29
Ok this is a little subtle - a column of `=`s doesn't work, but a column of `?`s is also incorrect because other base constructors should be considered equal to the dummy constructor.