@NathanielB123 2025-01-31
> E.g. how does the termination checking time for `Issue1075` change? (I suppose `--profile=internal` is the Agda option needed to display the accounting of runtime.)

On my laptop (R5 5600H), typechecking `Issue1075.agda` on this branch seems to have gotten slightly _faster_

(This branch)
```
Checking Issue1075 (/home/nathaniel/agda/fork/agda/test/Succeed/Issue1075.agda).
Total                        1,200ms        
Miscellaneous                   28ms        
Coverage                       118ms (435ms)
Coverage.UnifyIndices          317ms        
Typing                          46ms (377ms)
Typing.CheckLHS                 75ms (175ms)
Typing.CheckLHS.UnifyIndices   100ms        
Typing.CheckRHS                 99ms        
Typing.TypeSig                  41ms        
Termination                      8ms (123ms)
Termination.Graph              107ms        
Parsing                         35ms  (96ms)
Parsing.OperatorsExpr           49ms        
Parsing.OperatorsPattern        11ms        
Deserialization                 65ms  (73ms)
Scoping                         13ms  (35ms)
Scoping.InverseScopeLookup      21ms        
Highlighting                    28ms        
Positivity                      18ms        
InterfaceInstantiateFull        11ms
```
vs

(Master)
```
Checking Issue1075 (/home/nathaniel/agda/fork/agda/test/Succeed/Issue1075.agda).
Total                        1,278ms        
Miscellaneous                   30ms        
Coverage                       121ms (456ms)
Coverage.UnifyIndices          335ms        
Typing                          58ms (395ms)
Typing.CheckLHS                 77ms (181ms)
Typing.CheckLHS.UnifyIndices   103ms        
Typing.CheckRHS                102ms        
Typing.TypeSig                  39ms        
Termination                     18ms (178ms)
Termination.Graph              144ms        
Termination.Compare             14ms        
Parsing                         36ms  (96ms)
Parsing.OperatorsExpr           49ms        
Parsing.OperatorsPattern        11ms        
Deserialization                 65ms  (73ms)
Scoping                         14ms  (26ms)
Scoping.InverseScopeLookup      12ms        
Highlighting                    24ms        
Positivity                      19ms        
/home/nathaniel/agda/fork/agda/test/Succeed/Issue1075.agda:287.1-439.54: error: [TerminationIssue]
Termination checking failed for the following functions:
...
```
(both built with `--optimise-heavily`)

but I assume given the test succeeds on the new branch but fails previously, I'm not sure it's entirely fair. Compiling the standard library gives:

(This branch)
```
Total                                   449,879ms            
Miscellaneous                             2,711ms            
Typing                                   26,711ms (173,670ms)
Typing.CheckRHS                          54,052ms            
Typing.CheckLHS                          13,529ms  (36,293ms)
Typing.CheckLHS.UnifyIndices             22,763ms            
Typing.Generalize                        21,885ms            
Typing.OccursCheck                       14,189ms            
Typing.TypeSig                           13,966ms            
Typing.With                               5,667ms            
Typing.InstanceSearch                       140ms     (903ms)
Typing.InstanceSearch.FilterCandidates      491ms            
Typing.InstanceSearch.InitialCandidates     271ms            
Serialization                            54,175ms  (69,440ms)
Serialization.BinaryEncode                6,812ms            
Serialization.Compress                    5,970ms            
Serialization.Sort                        1,986ms            
Serialization.BuildInterface                495ms            
Coverage                                 34,240ms  (57,910ms)
Coverage.UnifyIndices                    23,670ms            
Scoping                                   5,033ms  (32,186ms)
Scoping.InverseScopeLookup               27,153ms            
Parsing                                   7,001ms  (23,986ms)
Parsing.OperatorsExpr                    14,139ms            
Parsing.OperatorsPattern                  2,845ms            
Deserialization                          15,155ms  (22,645ms)
Deserialization.Compaction                7,490ms            
Positivity                               21,027ms            
InterfaceInstantiateFull                 18,699ms            
DeadCode                                    332ms  (15,336ms)
DeadCode.DeadCodeReachable               15,003ms            
Termination                               1,592ms   (6,003ms)
Termination.RecCheck                      3,372ms            
Termination.Graph                           582ms            
Termination.Compare                         455ms            
Highlighting                              3,861ms            
Import                                    1,239ms            
Injectivity                                 867ms            
ProjectionLikeness                          242ms            
ModuleName                                   48ms
```
vs

(Master)
```
Total                                   414,393ms            
Miscellaneous                             2,612ms            
Typing                                   25,084ms (161,308ms)
Typing.CheckRHS                          50,574ms            
Typing.CheckLHS                          12,420ms  (33,858ms)
Typing.CheckLHS.UnifyIndices             21,437ms            
Typing.Generalize                        19,918ms            
Typing.OccursCheck                       13,026ms            
Typing.TypeSig                           12,905ms            
Typing.With                               5,098ms            
Typing.InstanceSearch                       132ms     (843ms)
Typing.InstanceSearch.FilterCandidates      464ms            
Typing.InstanceSearch.InitialCandidates     245ms            
Serialization                            50,331ms  (65,086ms)
Serialization.BinaryEncode                6,319ms            
Serialization.Compress                    5,720ms            
Serialization.Sort                        2,231ms            
Serialization.BuildInterface                483ms            
Coverage                                 31,879ms  (53,717ms)
Coverage.UnifyIndices                    21,838ms            
Scoping                                   4,794ms  (27,414ms)
Scoping.InverseScopeLookup               22,619ms            
Parsing                                   6,557ms  (22,299ms)
Parsing.OperatorsExpr                    13,131ms            
Parsing.OperatorsPattern                  2,610ms            
Deserialization                          14,216ms  (21,184ms)
Deserialization.Compaction                6,967ms            
Positivity                               19,822ms            
InterfaceInstantiateFull                 15,905ms            
DeadCode                                    319ms  (13,888ms)
DeadCode.DeadCodeReachable               13,568ms            
Termination                               1,613ms   (5,381ms)
Termination.RecCheck                      3,110ms            
Termination.Compare                         419ms            
Termination.Graph                           237ms            
Highlighting                              3,591ms            
Import                                    1,133ms            
Injectivity                                 780ms            
ProjectionLikeness                          224ms            
ModuleName                                   41ms
```
So apparently slower, but `Termination` is very similar, so maybe just random noise?

> Also, the comment on top of `Issue1075.agda` needs a new entry. Currently it says:

Ah yep good point - added