@NathanielB123 2025-02-01
Annoyingly I'm getting pretty inconsistent results, perhaps due to on other apps running in the background. After closing everything but the terminal running Agda (through WSL) I got:

This branch:
```
Total                                   399,772ms            
Miscellaneous                             2,507ms            
Typing                                   24,202ms (154,827ms)
Typing.CheckRHS                          47,767ms            
Typing.CheckLHS                          11,936ms  (32,361ms)
Typing.CheckLHS.UnifyIndices             20,424ms            
Typing.Generalize                        19,342ms            
Typing.OccursCheck                       12,960ms            
Typing.TypeSig                           12,479ms            
Typing.With                               4,864ms            
Typing.InstanceSearch                       124ms     (849ms)
Typing.InstanceSearch.FilterCandidates      442ms            
Typing.InstanceSearch.InitialCandidates     282ms            
Serialization                            47,659ms  (61,627ms)
Serialization.BinaryEncode                6,393ms            
Serialization.Compress                    5,613ms            
Serialization.Sort                        1,500ms            
Serialization.BuildInterface                461ms            
Coverage                                 30,554ms  (51,763ms)
Coverage.UnifyIndices                    21,209ms            
Scoping                                   4,484ms  (26,604ms)
Scoping.InverseScopeLookup               22,120ms            
Parsing                                   6,023ms  (21,099ms)
Parsing.OperatorsExpr                    12,569ms            
Parsing.OperatorsPattern                  2,506ms            
Deserialization                          13,860ms  (20,556ms)
Deserialization.Compaction                6,695ms            
Positivity                               19,532ms            
InterfaceInstantiateFull                 17,004ms            
DeadCode                                    307ms  (13,085ms)
DeadCode.DeadCodeReachable               12,778ms            
Termination                               1,394ms   (5,601ms)
Termination.RecCheck                      3,238ms            
Termination.Graph                           551ms            
Termination.Compare                         417ms            
Highlighting                              3,412ms            
Import                                    1,123ms            
Injectivity                                 761ms            
ProjectionLikeness                          221ms            
ModuleName                                   40ms
```

Master branch:
```
Total                                   407,501ms            
Miscellaneous                             2,569ms            
Typing                                   25,365ms (159,105ms)
Typing.CheckRHS                          49,031ms            
Typing.CheckLHS                          12,390ms  (33,480ms)
Typing.CheckLHS.UnifyIndices             21,090ms            
Typing.Generalize                        19,687ms            
Typing.OccursCheck                       13,108ms            
Typing.TypeSig                           12,515ms            
Typing.With                               5,097ms            
Typing.InstanceSearch                       128ms     (818ms)
Typing.InstanceSearch.FilterCandidates      453ms            
Typing.InstanceSearch.InitialCandidates     236ms            
Serialization                            49,003ms  (62,541ms)
Serialization.BinaryEncode                5,908ms            
Serialization.Compress                    5,648ms            
Serialization.Sort                        1,534ms            
Serialization.BuildInterface                447ms            
Coverage                                 31,147ms  (52,602ms)
Coverage.UnifyIndices                    21,454ms            
Scoping                                   4,692ms  (26,612ms)
Scoping.InverseScopeLookup               21,920ms            
Parsing                                   6,208ms  (21,796ms)
Parsing.OperatorsExpr                    12,997ms            
Parsing.OperatorsPattern                  2,590ms            
Deserialization                          13,937ms  (20,690ms)
Deserialization.Compaction                6,753ms            
Positivity                               20,305ms            
InterfaceInstantiateFull                 16,912ms            
DeadCode                                    327ms  (13,479ms)
DeadCode.DeadCodeReachable               13,151ms            
Termination                               1,523ms   (5,177ms)
Termination.RecCheck                      3,012ms            
Termination.Compare                         409ms            
Termination.Graph                           232ms            
Highlighting                              3,501ms            
Import                                    1,169ms            
Injectivity                                 772ms            
ProjectionLikeness                          221ms            
ModuleName                                   43ms
```

Not sure this enough data for firm conclusions, but it does appear that, on this branch, when typechecking the standard library, `Termination.Graph` pretty reliably takes longer (from 200-250ms to 550-600ms) and overall termination-checking time also increases (but by less proportionally), and overall typechecking time is basically unaffected (because termination checking is such a tiny fraction). I'll leave the decision of whether the trade-off makes sense to you (idk whether there are any scenarios where `Termination` actually starts to dominate runtime). I'm also happy to keep repeating the benchmark if you want further confirmation.